Remove unused htmltojsx
We no longer use htmltojsx to convert HTML to JSX and render within MDX.
Instead, we set HTML directly in a React component.

This package also did not work without patching on macOS due to line
ending issues so if we wanted to keep it, it would be ideal to commit a
patch to work cross-platform for development.