Memoize HTML so that details stay open on render
Without memoization, a change to the URL causes the router to re-render
the page. Clicking on a section's hash-link adds that section's prop ID
to the URL hash, causing a re-render. If another details element's
section was already open, the re-render of the page would cause the
other details elements to close. Memoization prevents excess closing of
details elements.